Authentication and authorization using Api's

+++ controllers/api/v1/post_api.js
@@ -31,7 +31,8 @@ module.exports.destroy= async function(req,res)
         let post= await Post.findById(req.params.id);
         
 
-       // if(post.user==req.user.id)
+        if(post.user==req.user.id)
+        {
        
             post.remove();
             await Comment.deleteMany({post:req.params.id});
@@ -54,11 +55,13 @@ module.exports.destroy= async function(req,res)
                 message:"Post and associated comments deleted Successfully"
                 
             });
+        }
         
-        // else{
-        //     req.flash('error',"you cannot delete this post!");
-        //     return res.redirect('back');
-        // }
+        else{
+            return res.json(401,{
+                message:"you cannot delete this post!"
+            })
+        }
         
     } catch (error) {
         console.log("error",error);
